% Copyright 2003 by Roger S. Bivand

\name{amacrines}
\title{Amacrines on/off data set}
\usage{data(amacrines)}
\alias{amacrines.on}
\alias{amacrines.off}
\format{Two two-column matrices of points marked on and off}
\source{\url{https://www.maths.lancs.ac.uk/~diggle/pointpatterns/Datasets/}, Peter J. Diggle, Department of Mathematics and Statistics, Lancaster University, Lancaster LA1 4YF, UK: public-domain spatial point pattern data-sets. 
}
\description{Two two-column matrices of points marked on and off}
\keyword{datasets}
