\name{area}
\alias{area}
\title{
Polygon Operations
}
\description{
Calculate area of polygon
}
\usage{
area(poly)
}
\arguments{
\item{poly}{
a polygon data set
}}
\value{
The area of the polygon is returned


}
\examples{
lancs.area_area(lancs.poly) # find area of lancaster boundary
}
% Converted by Sd2Rd version 0.3-3.
