\name{bodmin}
\title{Bodmin Moors granite tors}
\usage{data(bodmin)}
\alias{bodmin}
\format{A list corresponding to a Venables and Ripley point object with 35 observations
  \tabular{lll}{
    \tab x  \tab numeric  \tab grid eastings\cr
    \tab y   \tab numeric  \tab grid northings\cr
    \tab area \tab	list	\tab bounding box with xl, xu, yl, yu\cr
    \tab poly \tab array \tab polygon boundary with columns x and y\cr
  }
}
\source{Pinder and Witherick, 1977 - Bailey and Gatrell 1995, ch. 3.
}
\description{Locations of 35 granite tors on Bodmin Moor, taken from Infomap data set (northings multiplied by -1 to correspond to Figure 3.2, p. 82, Bailey and Gatrell.
}
\references{Bailey and Gatrell 1995, ch. 3.

}
\keyword{datasets}
