\name{cardiff}
\title{Locations of homes of juvenile offenders}
\usage{data(cardiff)}
\alias{cardiff}
\format{A Venables and Ripley point object with 168 observations
  \tabular{rlll}{
    [,1]  \tab Eastings  \tab numeric  \tab grid eastings\cr
    [,2]  \tab Northings   \tab numeric  \tab grid northings\cr
  }
}
\source{Herbert, 1980, - Bailey and Gatrell 1995, ch. 3.
}
\description{Locations of homes of 168 juvenile offenders on a Cardiff housing estate

}
\references{Bailey and Gatrell 1995, ch. 3.

}
\keyword{datasets}
