\name{cells}
\title{Centres of cells in tissue section}
\usage{data(cells)}
\alias{cells}
\format{A Venables and Ripley point object with 42 observations
  \tabular{rlll}{
    [,1]  \tab Eastings  \tab numeric  \tab grid eastings\cr
    [,2]  \tab Northings   \tab numeric  \tab grid northings\cr
  }
}
\source{Diggle 1983, - Bailey and Gatrell 1995, ch. 3.
}
\description{Locations of the centres of cells in a section of tissue

}
\references{Bailey and Gatrell 1995, ch. 3.

}
\keyword{datasets}
