\name{inout}
\alias{inout}
\title{
Point/Poly Operations
}
\description{
Test points for inclusion in a polygon.
}
\usage{
inout(pts,poly)
}
\arguments{
\item{pts}{
A points data set
}
\item{poly}{
A polygon data set
}}
\value{
A vector of logical values. \fCTrue} means the point was inside the
polygon, \fCFalse} means the point was outside.


}
\seealso{
\fCinpip},\fCpip}
}
% Converted by Sd2Rd version 0.3-3.


