\name{is.points}
\alias{is.points}
\title{
Point Objects
}
\description{
Tests for data in spatial point format.
}
\usage{
is.points(p)
}
\arguments{
\item{p}{
any S-PLUS object(s).
}
}
\value{
\code{is.points} returns TRUE if \code{p} is a points object, FALSE otherwise.
}
% Converted by Sd2Rd version 0.3-3.
