\name{larynx}
\title{Cases of cancer of larynx}
\usage{data(larynx)}
\alias{larynx}
\format{A Venables and Ripley point object with 57 observations
  \tabular{rlll}{
    [,1]  \tab Eastings  \tab numeric  \tab grid eastings\cr
    [,2]  \tab Northings   \tab numeric  \tab grid northings\cr
  }
}
\source{Diggle, Gatrell and Lovett, 1990, - Bailey and Gatrell 1995, ch. 3.
}
\description{Locations of cases of cancer of larynx in part of Lancashire

}
\references{Bailey and Gatrell 1995, ch. 3.

}
\keyword{datasets}
