\name{lung}
\title{Cases of cancer of lung}
\usage{data(lung)}
\alias{lung}
\format{A Venables and Ripley point object with 917 observations
  \tabular{rlll}{
    [,1]  \tab Eastings  \tab numeric  \tab grid eastings\cr
    [,2]  \tab Northings   \tab numeric  \tab grid northings\cr
  }
}
\source{Diggle, Gatrell and Lovett, 1990, - Bailey and Gatrell 1995, ch. 3.
}
\description{Locations of cases of cancer of lung in part of Lancashire

}
\references{Bailey and Gatrell 1995, ch. 3.

}
\keyword{datasets}
