\name{nndistG}
\alias{nndistG}
\title{
Point Summary Statistics
}
\description{
Calculate nearest neighbour distances as used by \fCGhat()}.
}
\usage{
nndistG(pts)
}
\arguments{
\item{pts}{
A points data set
}}
\value{
Returns a list with components \fC$dists} and \fC$neighs}.
\fC$dists[i]} is the distance
to the nearest neighbour of point \fCi} in \fCpts,} and \fC$neighs[i]}
is the index
of the neighbour of point \fCi}.


}
\seealso{
\fCnndistF},\fCFhat},\fCGhat},\fCFzero}
}
% Converted by Sd2Rd version 0.3-3.
