\name{okwhite}
\title{Oklahoma white offenders}
\usage{data(okwhite)}
\alias{okwhite}
\format{A list corresponding to a Venables and Ripley point object with 104 observations
  \tabular{lll}{
    x  \tab numeric  \tab grid eastings\cr
    y   \tab numeric  \tab grid northings\cr
    area \tab	list	\tab bounding box with xl, xu, yl, yu\cr
  }
}
\source{Carter and Hill, 1979, - Bailey and Gatrell 1995, ch. 3.
}
\description{Locations of theft from property offences committed by white offenders in Oklahoma City

}
\references{Bailey and Gatrell 1995, ch. 3.

}
\keyword{datasets}
