\name{redwood}
\title{Redwood seedlings in a forest}
\usage{data(redwood)}
\alias{redwood}
\format{A Venables and Ripley point object with 62 observations
  \tabular{rlll}{
    [,1]  \tab Eastings  \tab numeric  \tab grid eastings\cr
    [,2]  \tab Northings   \tab numeric  \tab grid northings\cr
  }
}
\source{Diggle, 1983, - Bailey and Gatrell 1995, ch. 3.
}
\description{Locations of redwood seedlings in a forest

}
\references{Bailey and Gatrell 1995, ch. 3.

}
\keyword{datasets}
