\name{rlabel}
\alias{rlabel}
\title{
Point Operations
}
\description{
Randomly label two or more point sets.
}
\usage{
rlabel(\dots)
}
\arguments{
\item{...}{
Any number of points data sets
}}
\value{
A list of points data sets. There are as many elements in the list as 
arguments.


}
\section{METHOD}{
The output data sets are a random labelling of the input data sets, i.e. 
all the points in the input data sets are randomly assigned to the output 
sets. The number of points in each output set is the same as its corresponding
input set.
}
% Converted by Sd2Rd version 0.3-3.
