\name{sbox}
\alias{sbox}
\title{
Generate a box surrounding a point object
}
\description{
Generate a box surrounding a point object
}
\usage{
sbox(pts, xfrac = .1, yfrac = .1)
}
\arguments{
\item{pts}{
A points data set
}
\item{xfrac}{
The fraction of the width of the point pattern by which the box will
surround the point pattern to the left and right.
}
\item{yfrac}{
The fraction of the height of the point pattern by which the box will 
surround the point pattern to the top and bottom. 
}}
\value{
A points data set of four points giving the coordinates of the surrounding box


}
\seealso{
\code{bbox}
}
% Converted by Sd2Rd version 0.3-3.
