\name{shift}
\alias{shift}
\title{
Shift a point data set
}
\description{
Shift a point data set
}
\usage{
shift(pts,xsh=0.0,ysh=0.0)
}
\arguments{
\item{pts}{
The point data set to shift
}
\item{xsh}{
Amount to shift along the x-axis
}
\item{ysh}{
Amount to shift along the y-axis
}}
\value{
A point data set like \code{pts}, but with \code{xsh} added to its
x-coordinates, and \code{ysh} added to its y-coordinates.


}
\seealso{
\code{tor.shift}
}
% Converted by Sd2Rd version 0.3-3.
