% Copyright Barry Rowlingson <b.rowlingson@lancaster.ac.uk> and 
% Peter Diggle (c) 1991-3; http://www.maths.lancs.ac.uk/~rowlings/Splancs/
% R port: copyright 1998-2000 by Roger S. Bivand
%
%  This program is free software; you can redistribute it and/or modify
%  it under the terms of the GNU General Public License as published by
%  the Free Software Foundation; either version 2 of the License, or
%  (at your option) any later version.
%
%  This program is distributed in the hope that it will be useful,
%  but WITHOUT ANY WARRANTY; without even the implied warranty of
%  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
%  GNU General Public License for more details.
%

\name{splancs}
\alias{splancs}
\title{
Return version number and author information
}
\description{
Return version number and author information
}
\usage{
splancs()
}
\value{
The version string is returned. This is a number of the format x.yy,
where x is the major version number and yy is the minor version number.


}
\section{Side Effects}{
A message is printed giving copyright and author information, together with
the date of release of the software. This information is useful in case of
problems with Splancs.


}
% Converted by Sd2Rd version 0.3-3.
