\name{thin}
\alias{thin}
\title{
Randomly thin a point data set
}
\description{
Randomly thin a point data set.
}
\usage{
thin(pts,n)
}
\arguments{
\item{pts}{
a points data set.
}
\item{n}{
the number of points to return
}}
\value{
Returns a point data set consisting of \code{n} points selected randomly
from the set \code{pts}.
}
% Converted by Sd2Rd version 0.3-3.
