\name{uganda}
\title{Craters in Uganda}
\usage{data(uganda)}
\alias{uganda}
\format{A Venables and Ripley point object with 120 observations
  \tabular{rlll}{
    [,1]  \tab Eastings  \tab numeric  \tab grid eastings\cr
    [,2]  \tab Northings   \tab numeric  \tab grid northings\cr
  }
}
\source{Tinkler, 1971, - Bailey and Gatrell 1995, ch. 3.
}
\description{Locations of craters in a volcanic field in Uganda

}
\references{Bailey and Gatrell 1995, ch. 3.

}
\keyword{datasets}
