\name{bbox}
\alias{bbox}
\title{
Polygon Operations
}
\description{
Generate a bounding box round a point object
}
\usage{
bbox(pts)
}
\arguments{
\item{pts}{
A points data set
}}
\value{
A points data set of four points giving the coordinates of the bounding box


}
\seealso{
\code{sbox}
}
% Converted by Sd2Rd version 0.3-3.
