\name{bodmin}
\title{Bodmin Moors granite tors}
\usage{data(bodmin)}
\alias{bodmin}
\format{A Venables and Ripley point object with 35 observations
  \tabular{rlll}{
    [,1]  \tab Eastings  \tab numeric  \tab grid eastings\cr
    [,2]  \tab Northings   \tab numeric  \tab grid northings\cr
  }
}
\source{Pinder and Witherick, 1977 - Bailey and Gatrell 1995, ch. 3.
}
\description{Locations of 35 granite tors on Bodmin Moor

}
\references{Bailey and Gatrell 1995, ch. 3.

}
\keyword{datasets}
