\name{cardiff}
\title{Locations of homes of juvenile offenders}
\usage{data(cardiff)}
\alias{cardiff}
\format{A list corresponding to a Venables and Ripley point object with 168 observations
  \tabular{lll}{
    x  \tab numeric  \tab grid eastings\cr
    y   \tab numeric  \tab grid northings\cr
    area \tab	list	\tab bounding box with xl, xu, yl, yu\cr
    poly \tab array \tab polygon boundary with columns x and y\cr
  }
}
\source{Herbert, 1980, - Bailey and Gatrell 1995, ch. 3.
}
\description{Locations of homes of 168 juvenile offenders on a Cardiff housing estate

}
\references{Bailey and Gatrell 1995, ch. 3.

}
\keyword{datasets}
