\name{inpip}
\alias{inpip}
\title{
Point/Poly Operations
}
\description{
Select points inside a polygon
}
\usage{
inpip(pts,poly)
}
\arguments{
\item{pts}{
A points data set
}
\item{poly}{
A polygon data set
}}
\value{
\code{inpip} returns a vector of indices of the points in \code{pts} that are located
in the polygon.


}
\seealso{
\code{pip}
}
% Converted by Sd2Rd version 0.3-3.
