% Copyright Barry Rowlingson <b.rowlingson@lancaster.ac.uk> and 
% Peter Diggle (c) 1991-3; http://www.maths.lancs.ac.uk/~rowlings/Splancs/
% R port: copyright 1998-2000 by Roger S. Bivand
%
%  This program is free software; you can redistribute it and/or modify
%  it under the terms of the GNU General Public License as published by
%  the Free Software Foundation; either version 2 of the License, or
%  (at your option) any later version.
%
%  This program is distributed in the hope that it will be useful,
%  but WITHOUT ANY WARRANTY; without even the implied warranty of
%  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
%  GNU General Public License for more details.
%

\name{is.points}
\alias{is.points}
\title{
Point Objects
}
\description{
Tests for data in spatial point format.
}
\usage{
is.points(p)
}
\arguments{
\item{p}{
any S-PLUS object(s).
}
}
\value{
\code{is.points} returns TRUE if \code{p} is a points object, FALSE otherwise.
}
% Converted by Sd2Rd version 0.3-3.
