\name{nndistF}
\alias{nndistF}
\title{
Point Summary Statistics
}
\description{
Calculate nearest neighbour distances as used by  \fCFhat()}
}
\usage{
nndistF(pts1,pts2)
}
\arguments{
\item{pts1}{
A points data set
}
\item{pts2}{
A points data set
}}
\value{
The set of distances from each of the points in \fCpts2} to the nearest
point in \fCpts1} is returned as a vector.


}
\seealso{
\fCnndistG},\fCFhat},\fCGhat},\fCFzero}
}
% Converted by Sd2Rd version 0.3-3.
