\name{okblack}
\title{Oklahoma black offenders}
\usage{data(okblack)}
\alias{okblack}
\format{A Venables and Ripley point object with 147 observations
  \tabular{rlll}{
    [,1]  \tab Eastings  \tab numeric  \tab grid eastings\cr
    [,2]  \tab Northings   \tab numeric  \tab grid northings\cr
  }
}
\source{Carter and Hill, 1979, - Bailey and Gatrell 1995, ch. 3.
}
\description{Locations of theft from property offences committed by black offenders in Oklahoma City

}
\references{Bailey and Gatrell 1995, ch. 3.

}
\keyword{datasets}
