\name{okwhite}
\title{Oklahoma white offenders}
\usage{data(okwhite)}
\alias{okwhite}
\format{A Venables and Ripley point object with 104 observations
  \tabular{rlll}{
    [,1]  \tab Eastings  \tab numeric  \tab grid eastings\cr
    [,2]  \tab Northings   \tab numeric  \tab grid northings\cr
  }
}
\source{Carter and Hill, 1979, - Bailey and Gatrell 1995, ch. 3.
}
\description{Locations of theft from property offences committed by white offenders in Oklahoma City

}
\references{Bailey and Gatrell 1995, ch. 3.

}
\keyword{datasets}
