% Copyright Barry Rowlingson  and 
% Peter Diggle (c) 1991-3; http://www.maths.lancs.ac.uk/~rowlings/Splancs/
% R port: copyright 1998-2000 by Roger S. Bivand
%
%  This program is free software; you can redistribute it and/or modify
%  it under the terms of the GNU General Public License as published by
%  the Free Software Foundation; either version 2 of the License, or
%  (at your option) any later version.
%
%  This program is distributed in the hope that it will be useful,
%  but WITHOUT ANY WARRANTY; without even the implied warranty of
%  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
%  GNU General Public License for more details.
%

\name{okwhite}
\title{Oklahoma white offenders}
\usage{data(okwhite)}
\alias{okwhite}
\format{A list corresponding to a Venables and Ripley point object with 104 observations
  \tabular{lll}{
    x  \tab numeric  \tab grid eastings\cr
    y   \tab numeric  \tab grid northings\cr
    area \tab	list	\tab bounding box with xl, xu, yl, yu\cr
  }
}
\source{Carter and Hill, 1979, - Bailey and Gatrell 1995, ch. 3.
}
\description{Locations of theft from property offences committed by white offenders in Oklahoma City

}
\references{Bailey, T. C. and Gatrell, A. C. 1995, Interactive spatial data analysis. Longman, Harlow.

}
\keyword{datasets}
