\name{pdense}
\alias{pdense}
\title{
Overall density for a point pattern
}
\description{
Calculate overall density for a point pattern.
}
\usage{
pdense(pts,poly)
}
\arguments{
\item{pts}{
A points data set
}
\item{poly}{
A polygon data set
}}
\value{
The density of the points in the polygon. i.e. the number of points
per unit area.


}
\seealso{
\code{Fzero}
}
% Converted by Sd2Rd version 0.3-3.
