\name{splancs}
\alias{splancs}
\title{
Point pattern analysis
}
\description{
Return version number and author information
}
\usage{
splancs()
}
\value{
The version string is returned. This is a number of the format x.yy,
where x is the major version number and yy is the minor version number.


}
\section{Side Effects}{
A message is printed giving copyright and author information, together with
the date of release of the software. This information is useful in case of
problems with Splancs.


}
% Converted by Sd2Rd version 0.3-3.
