\name{uganda}
\title{Craters in Uganda}
\usage{data(uganda)}
\alias{uganda}
\format{A list corresponding to a Venables and Ripley point object with 120 observations
  \tabular{lll}{
    x  \tab numeric  \tab grid eastings\cr
    y   \tab numeric  \tab grid northings\cr
    area \tab	list	\tab bounding box with xl, xu, yl, yu\cr
    poly \tab array \tab polygon boundary with columns x and y\cr
  }
}
\source{Tinkler, 1971, - Bailey and Gatrell 1995, ch. 3.
}
\description{Locations of craters in a volcanic field in Uganda

}
\references{Bailey and Gatrell 1995, ch. 3.

}
\keyword{datasets}
